Int64

+++ src/mcmc_algorithm.jl
@@ -191,7 +191,7 @@ MCMCSpec(
 
 
 function (spec::MCMCSpec)(
-    id::Int,
+    id::Int64,
     exec_context::ExecContext = ExecContext()
 )
     P = float(eltype(param_bounds(spec.prior)))
